Working LoginFragment

+++ app/src/main/java/net/iessochoa/grupof/practicafinalandroid/model/Song.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "songs")
 class Song(
-    @PrimaryKey(autoGenerate = true) val id: Float,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     val name: String,
     val artist: String,
     val album: String,
